// src/app/device-only/ais/device-only-ais/models/product.model.ts
export class Product {
    status: string;
    message: string;
    brand: string;
    model: string;
    color: string;
    company: string;
    matcode: string;
    invNo: string;
    invDt: string;
    invDealerCode: string;
    invItemCat: string;
    commercialName: string;
    productType?: string;
    productSubtype?: string;
}